projects/composable-features/pyproject.toml
[tool.poetry]
name = "composable-features"
version = "0.0.1rc0"
description = "Composable feature spaces for multi-task learning."
authors = ["Anderson de Andrade <anderson_de_andrade@sfu.ca>"]
exclude = ["tests/*"]

[[tool.poetry.source]]
name = "private"
url = "https://pypi.org/simple"
secondary = true

[tool.poetry.dependencies]
python = "3.9.9"
annoy = "1.17.0"
compressai = "1.2.0"
defopt = "6.1.0"
matplotlib = "3.4.3"
mlflow-skinny = "1.24.0"
numba = "0.55.1"
numpy = "1.21.5"
pillow = "9.0.1"
pycocotools = "2.0.3"
pytorch-lightning = "1.5.10"
scipy = "1.7.3"
seaborn = "0.11.2"
sfu-torch-lib = "0.0.11"
torch = "1.11.0"

[tool.poetry.extras]
extras = []

[tool.poetry.dev-dependencies]
defopt = "*"
flake8 = "*"
mypy = "*"
poethepoet = "*"
pytest = "*"

[tool.poe.tasks]
cu113 = "python -m pip install torch==1.11.0+cu113 torchvision==0.12.0+cu113 --no-deps --upgrade --find-links https://download.pytorch.org/whl/torch_stable.html"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
